+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - added config option to configure a webroot for the ui
 - prevent url-based settings from being changed from a pwa
 - added config option to for inactivity-based session timeout
+- add search bar on settings page for filtering options
 
 ### Other Changes